requirements.txt
chromadb>=0.4.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pytest>=7.0.0
